+++ rpc/ums/internal/logic/memberlevelservice/memberlevelupdatelogic.go
@@ -33,7 +33,7 @@ func (l *MemberLevelUpdateLogic) MemberLevelUpdate(in *umsclient.MemberLevelUpda
 		LevelName:          in.LevelName,
 		GrowthPoint:        in.GrowthPoint,
 		DefaultStatus:      in.DefaultStatus,
-		FreeFreightPoint:   float64(in.FreeFreightPoint),
+		FreeFreightPoint:   in.FreeFreightPoint,
 		CommentGrowthPoint: in.CommentGrowthPoint,
 		IsFreeFreight:      in.IsFreeFreight,
 		IsSignIn:           in.IsSignIn,
